[Rename] - Stripable => Strippable.
