+++ libse/LibSE.csproj
@@ -190,7 +190,7 @@
     <Compile Include="SpellCheck\UndoObject.cs" />
     <Compile Include="SsaStyle.cs" />
     <Compile Include="StringExtensions.cs" />
-    <Compile Include="StripableText.cs" />
+    <Compile Include="StrippableText.cs" />
     <Compile Include="SubtitleEditRegex.cs" />
     <Compile Include="Subtitle.cs" />
     <Compile Include="SubtitleFormats\AribB36.cs" />
