+++ src/Test/Logic/StrippableTextTest.cs
@@ -4,112 +4,112 @@ using Nikse.SubtitleEdit.Core;
 namespace Test.Logic
 {
     [TestClass]
-    public class StripableTextTest
+    public class StrippableTextTest
     {
 
         [TestMethod]
-        public void StripableTextItalic()
+        public void StrippableTextItalic()
         {
-            var st = new StripableText("<i>Hi!</i>");
+            var st = new StrippableText("<i>Hi!</i>");
             Assert.AreEqual(st.Pre, "<i>");
             Assert.AreEqual(st.Post, "!</i>");
             Assert.AreEqual(st.StrippedText, "Hi");
         }
 
         [TestMethod]
-        public void StripableTextAss()
+        public void StrippableTextAss()
         {
-            var st = new StripableText("{\\an9}Hi!");
+            var st = new StrippableText("{\\an9}Hi!");
             Assert.AreEqual(st.Pre, "{\\an9}");
             Assert.AreEqual(st.Post, "!");
             Assert.AreEqual(st.StrippedText, "Hi");
         }
 
         [TestMethod]
-        public void StripableTextFont()
+        public void StrippableTextFont()
         {
-            var st = new StripableText("<font color=\"red\">Hi!</font>");
+            var st = new StrippableText("<font color=\"red\">Hi!</font>");
             Assert.AreEqual(st.Pre, "<font color=\"red\">");
             Assert.AreEqual(st.Post, "!</font>");
             Assert.AreEqual(st.StrippedText, "Hi");
         }
 
         [TestMethod]
-        public void StripableTextItalic2()
+        public void StrippableTextItalic2()
         {
-            var st = new StripableText("<i>O</i>");
+            var st = new StrippableText("<i>O</i>");
             Assert.AreEqual(st.Pre, "<i>");
             Assert.AreEqual(st.Post, "</i>");
             Assert.AreEqual(st.StrippedText, "O");
         }
 
         [TestMethod]
-        public void StripableTextItalic3()
+        public void StrippableTextItalic3()
         {
-            var st = new StripableText("<i>Hi!");
+            var st = new StrippableText("<i>Hi!");
             Assert.AreEqual(st.Pre, "<i>");
             Assert.AreEqual(st.Post, "!");
             Assert.AreEqual(st.StrippedText, "Hi");
         }
 
         [TestMethod]
-        public void StripableTextFontDontTouch()
+        public void StrippableTextFontDontTouch()
         {
-            var st = new StripableText("{MAN} Hi, how are you today!");
+            var st = new StrippableText("{MAN} Hi, how are you today!");
             Assert.AreEqual(st.Pre, "");
             Assert.AreEqual(st.Post, "!");
             Assert.AreEqual(st.StrippedText, "{MAN} Hi, how are you today");
         }
 
         [TestMethod]
-        public void StripableOnlyPre()
+        public void StrippableOnlyPre()
         {
-            var st = new StripableText("(");
+            var st = new StrippableText("(");
             Assert.AreEqual(st.Pre, "(");
             Assert.AreEqual(st.Post, "");
             Assert.AreEqual(st.StrippedText, "");
         }
 
         [TestMethod]
-        public void StripableOnlyPre2()
+        public void StrippableOnlyPre2()
         {
-            var st = new StripableText("<");
+            var st = new StrippableText("<");
             Assert.AreEqual(st.Pre, "");
             Assert.AreEqual(st.Post, "");
             Assert.AreEqual(st.StrippedText, "<");
         }
 
         [TestMethod]
-        public void StripableOnlyPre3()
+        public void StrippableOnlyPre3()
         {
-            var st = new StripableText("<i>");
+            var st = new StrippableText("<i>");
             Assert.AreEqual(st.Pre, "<i>");
             Assert.AreEqual(st.Post, "");
             Assert.AreEqual(st.StrippedText, "");
         }
 
         [TestMethod]
-        public void StripableOnlyText()
+        public void StrippableOnlyText()
         {
-            var st = new StripableText("H");
+            var st = new StrippableText("H");
             Assert.AreEqual(st.Pre, "");
             Assert.AreEqual(st.Post, "");
             Assert.AreEqual(st.StrippedText, "H");
         }
 
         [TestMethod]
-        public void StripableTextItalicAndFont()
+        public void StrippableTextItalicAndFont()
         {
-            var st = new StripableText("<i><font color=\"red\">Hi!</font></i>");
+            var st = new StrippableText("<i><font color=\"red\">Hi!</font></i>");
             Assert.AreEqual(st.Pre, "<i><font color=\"red\">");
             Assert.AreEqual(st.Post, "!</font></i>");
             Assert.AreEqual(st.StrippedText, "Hi");
         }
 
         [TestMethod]
-        public void StripableTextItalicAndMore()
+        public void StrippableTextItalicAndMore()
         {
-            var st = new StripableText("<i>...<b>Hi!</b></i>");
+            var st = new StrippableText("<i>...<b>Hi!</b></i>");
             Assert.AreEqual(st.Pre, "<i>...<b>");
             Assert.AreEqual(st.Post, "!</b></i>");
             Assert.AreEqual(st.StrippedText, "Hi");
